requirements.txt
gradio==3.50.2
spacy==3.7.2
typer==0.9.4
websockets==11.0.3
streamlit==1.32.2
transformers==4.39.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
nltk==3.8.1
python-dotenv==1.0.1
google-api-python-client==2.126.0
tensorflow==2.15.0

